Enable the message attachments saga for PN
